requirements.txt
flask
python-telegram-bot
reportlab
pillow
psycopg2-binary

